Replaced safe_reinterpret_cast & reinterpret_cast with new erasure_cast function.
erasure_cast overcomes a bug found in the MSVC++ 2017 compiler (up to and including MSVC++ 15.8 Preview 5) where the compiler was incorrectly aliasing the type-punned values in the safe_reinterpret_cast union under certain circumstances.  This resulted in a garbage value being stored into system_code causing the unit tests to fail.

erasure_cast is also guaranteed to be constexpr when the input and output types are either integral types or enum types, which contrasts with safe_reinterpret_cast and reinterpret_cast, neither of which support usage in compile-time expressions (at least in C++17 and prior). It is written in terms of a partially compliant C++20 std::bit_cast function to illustrate how this future library facility can be leveraged by implementors.