BREAKING CHANGE: Make status_code's explicit copy constructor from status_code<void> throw an exception if the erased copy fails. There is a new nothrow overload which preserves the old behaviour.
Also pass exclusive responsibility for determining whether an erased copy can occur or not to the source domain's erased copy implementation rather than testing locally if the destination status code is big enough. This enables domains to drop non-mandatory state if the destination is smaller than the source, but not too small.